+++ reNgine/settings.py
@@ -1,6 +1,6 @@
 import os
-
 from reNgine.init import first_run
+from celery.schedules import crontab
 
 
 # !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
